requirements.txt
python-binance
requests
pydantic
